Split BasicRippersTest.java into per-ripper test files.
