+++ src/test/java/com/rarchives/ripme/tst/AppTest.java
@@ -10,21 +10,21 @@ public class AppTest extends TestCase {
      *
      * @param testName name of the test case
      */
-    public AppTest( String testName ) {
-        super( testName );
+    public AppTest(String testName) {
+        super(testName);
     }
 
     /**
      * @return the suite of tests being tested
      */
     public static Test suite() {
-        return new TestSuite( AppTest.class );
+        return new TestSuite(AppTest.class);
     }
 
     /**
      * Rigourous Test :-)
      */
     public void testApp() {
-        assertTrue( true );
+        assertTrue(true);
     }
 }
