+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/FineboxRipperTest.java
@@ -1,0 +1,22 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.FineboxRipper;
+
+public class FineboxRipperTest extends RippersTest {
+    /*
+    public void testVineboxAlbum() throws IOException {
+        FineboxRipper ripper = new FineboxRipper(new URL("http://vinebox.co/u/wi57hMjc2Ka"));
+        testRipper(ripper);
+    }
+    */
+
+    /*
+    public void testFineboxAlbum() throws IOException {
+        FineboxRipper ripper = new FineboxRipper(new URL("http://finebox.co/u/wi57hMjc2Ka"));
+        testRipper(ripper);
+    }
+    */
+}
