+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/FivehundredpxRipperTest.java
@@ -1,0 +1,13 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.FivehundredpxRipper;
+
+public class FivehundredpxRipperTest extends RippersTest {
+    public void test500pxAlbum() throws IOException {
+        FivehundredpxRipper ripper = new FivehundredpxRipper(new URL("https://marketplace.500px.com/alexander_hurman"));
+        testRipper(ripper);
+    }
+}
