+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/FlickrRipperTest.java
@@ -1,0 +1,15 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.FlickrRipper;
+
+public class FlickrRipperTest extends RippersTest {
+    /*
+    public void testFlickrAlbum() throws IOException {
+        FlickrRipper ripper = new FlickrRipper(new URL("https://www.flickr.com/photos/leavingallbehind/sets/72157621895942720/"));
+        testRipper(ripper);
+    }
+    */
+}
