+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/ImagestashRipperTest.java
@@ -1,0 +1,15 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.ImagestashRipper;
+
+public class ImagestashRipperTest extends RippersTest {
+    /*
+    public void testImagestashRip() throws IOException {
+        ImagestashRipper ripper = new ImagestashRipper(new URL("https://imagestash.org/tag/everydayuncensor"));
+        testRipper(ripper);
+    }
+    */
+}
