+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/MotherlessRipperTest.java
@@ -1,0 +1,13 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.MotherlessRipper;
+
+public class MotherlessRipperTest extends RippersTest {
+    public void testMotherlessAlbumRip() throws IOException {
+        MotherlessRipper ripper = new MotherlessRipper(new URL("http://motherless.com/G4DAA18D"));
+        testRipper(ripper);
+    }
+}
