+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/RedditRipperTest.java
@@ -1,0 +1,24 @@
+
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.RedditRipper;
+
+public class RedditRipperTest extends RippersTest {
+    /*
+    public void testRedditSubredditRip() throws IOException {
+        RedditRipper ripper = new RedditRipper(new URL("http://www.reddit.com/r/nsfw_oc"));
+        testRipper(ripper);
+    }
+    public void testRedditSubredditTopRip() throws IOException {
+        RedditRipper ripper = new RedditRipper(new URL("http://www.reddit.com/r/nsfw_oc/top?t=all"));
+        testRipper(ripper);
+    }
+    public void testRedditPostRip() throws IOException {
+        RedditRipper ripper = new RedditRipper(new URL("http://www.reddit.com/r/UnrealGirls/comments/1ziuhl/in_class_veronique_popa/"));
+        testRipper(ripper);
+    }
+    */
+}
