+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/SankakuComplexRipperTest.java
@@ -1,0 +1,19 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.SankakuComplexRipper;
+
+public class SankakuComplexRipperTest extends RippersTest {
+    /*
+    public void testSankakuChanRip() throws IOException {
+        SankakuComplexRipper ripper = new SankakuComplexRipper(new URL("https://chan.sankakucomplex.com/?tags=cleavage"));
+        testRipper(ripper);
+    }
+    public void testSankakuIdolRip() throws IOException {
+        SankakuComplexRipper ripper = new SankakuComplexRipper(new URL("https://idol.sankakucomplex.com/?tags=meme_%28me%21me%21me%21%29_%28cosplay%29"));
+        testRipper(ripper);
+    }
+    */
+}
