+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/TwodgalleriesRipperTest.java
@@ -1,0 +1,16 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.TwodgalleriesRipper;
+
+public class TwodgalleriesRipperTest extends RippersTest {
+    /*
+    public void testTwodgalleriesRip() throws IOException {
+        TwodgalleriesRipper ripper = new TwodgalleriesRipper(new URL("http://www.2dgalleries.com/artist/regis-loisel-6477"));
+        testRipper(ripper);
+    }
+    */
+}
+
