+++ src/test/java/com/rarchives/ripme/tst/ripper/rippers/VineRipperTest.java
@@ -1,0 +1,15 @@
+package com.rarchives.ripme.tst.ripper.rippers;
+
+import java.io.IOException;
+import java.net.URL;
+
+import com.rarchives.ripme.ripper.rippers.VineRipper;
+
+public class VineRipperTest extends RippersTest {
+    /*
+    public void testVineRip() throws IOException {
+        VineRipper ripper = new VineRipper(new URL("https://vine.co/u/954440445776334848"));
+        testRipper(ripper);
+    }
+    */
+}
